bit stream begin

+++ liblgcodec/codec/common/i_packetizer.h
@@ -19,8 +19,8 @@ class IPacketizer
 {
 public:
 	virtual ~IPacketizer() {}
-	virtual std::int32_t Push(const void* data, std::size_t size) = 0;
-	virtual std::int32_t Pop(void* data, std::size_t size = 0) = 0;
+	virtual std::size_t Push(const void* data, std::size_t size) = 0;
+	virtual std::size_t Pop(void* data, std::size_t size = 0) = 0;
 	virtual void Reset() = 0;
 	virtual std::size_t Count() const = 0;
 	virtual packetizer_direction_t GetDirection() const = 0;
